+++ deltafi-common/src/main/java/org/deltafi/common/types/PluginCoordinates.java
@@ -50,6 +50,10 @@ public class PluginCoordinates {
     version = matcher.group(3);
   }
 
+  public String groupAndArtifact() {
+    return groupId + ":" + artifactId;
+  }
+
   public boolean equalsIgnoreVersion(PluginCoordinates pluginCoordinates) {
     return groupId.equals(pluginCoordinates.groupId) && artifactId.equals(pluginCoordinates.artifactId);
   }